+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### 功能构建
 
+- 实现 `familyhelper-life` 模块的推送器。
+  - com.dwarfeng.familyhelper.plugin.life.handler.pusher.FamilyhelperPusher。
+
 - 实现 `familyhelper-life` 模块的服务。
   - com.dwarfeng.familyhelper.plugin.life.service.DubboRestActivityFileOperateService。